requirements.txt
django-oauth-toolkit>=3.0.1
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-environ>=0.11.2